Fix steering accumulation bug: keep steering as float to avoid truncation
The bug: steering was int16_t, so delta_steering values < 1.0 were
truncated to 0 when cast to int32_t before accumulating.

With sensitivity=20, each mouse pixel only adds ±0.2, which became 0.

Solution: Store steering as float internally, only convert to int16_t
when sending to gamepad device. This allows fractional values to
accumulate properly.

+++ src/gamepad.h
@@ -30,7 +30,7 @@ private:
     int fd;
     
     // State
-    int16_t steering;
+    float steering;  // Keep as float to avoid truncation, convert to int16_t when sending
     float throttle;
     float brake;
     std::map<std::string, bool> buttons;
